Re{factor|package} M.S.I.Stream and VarlenaWrapper
There is nothing very JDBC-specific about MarkableSequenceInputStream,
so move it from o.p.p.jdbc to o.p.p.internal. VarlenaWrapper.Input is
mostly made up of what could be a more general ByteBufferInputStream,
so factor that out into its own class.
